Refactor: extracting exercises_submissions_helper
